# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Todo": {
				"properties": {
					"TodoId": {
						"type": "string"
					},
					"Title": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"Urgency": {
						"type": "string"
					},
					"DueDate": {
						"type": "string"
					},
					"CreatedAt": {
						"type": "string"
					}
				},
				"required": [
					"TodoId",
					"Title",
					"Description",
					"Urgency",
					"DueDate",
					"CreatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddTodoValidator": {
				"properties": {
					"Title": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"Urgency": {
						"type": "string"
					},
					"DueDate": {
						"type": "string"
					}
				},
				"required": [
					"Title",
					"Description",
					"Urgency",
					"DueDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTodoValidator": {
				"properties": {
					"Title": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"Urgency": {
						"type": "string"
					},
					"DueDate": {
						"type": "string"
					}
				},
				"required": [
					"Urgency"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/todos": {
			"get": {
				"operationId": "GetTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Todo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/todo/{TodoId}": {
			"get": {
				"operationId": "GetOneTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "TodoId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "TodoId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTodoValidator"
							}
						}
					}
				}
			}
		},
		"/todo": {
			"post": {
				"operationId": "AddTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Todo"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddTodoValidator"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}